Reorganize the Cpp files separate from the python library
