+++ src/core/add/README.md
@@ -1,0 +1,3 @@
+# ADD
+
+Just used for setting up the libseis package.
