+++ src/core/add/add.cpp
@@ -1,0 +1,9 @@
+//
+// Created by dg on 6/7/25.
+//
+
+#include "add.hpp"
+
+namespace libseis {
+int add(const int i, const int j) { return i + j; }
+} // namespace libseis
