+++ src/core/velocity/velocity.hpp
