+++ src/libseis/_num.pyi
