+++ src/libseis/math.py
@@ -1,0 +1,10 @@
+from ._num import add
+
+
+def add_one(i: int) -> int:
+    """
+    Add one to the given integer.
+
+    This function takes an integer and returns the integer incremented by one.
+    """
+    return add(i, 1)
